dia malo codigo peor

+++ src/main/java/com/proyectofinal/ecomerce/dao/CarritoDao.java
@@ -1,10 +1,15 @@
 package com.proyectofinal.ecomerce.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 
 import com.proyectofinal.ecomerce.model.Carrito;
+import com.proyectofinal.ecomerce.model.Producto;
 
-public interface CarritoDao  extends JpaRepository<Carrito, Integer>{
-	
+public interface CarritoDao extends JpaRepository<Carrito, Long> {
+
+	//@Query("")
+	//void addItem(Producto producto);
+
+	//void deleteItem(Producto producto);
 }
-
